% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitocarta2.hg19.R
\docType{data}
\name{mitocarta2.hg19}
\alias{mitocarta2.hg19}
\title{MitoCarta 2.0: an atlas of mitochondrial genes and proteins}
\format{A GRanges object.}
\source{
\url{http://www.broadinstitute.org/ftp/distribution/metabolic/papers/Pagliarini/MitoCarta2.0/Human.MitoCarta2.0.bed}
}
\usage{
mitocarta2.hg19
}
\description{
This is the hg19 (human) version of MitoCarta 2.0, downloaded from 
https://www.broadinstitute.org/scientific-community/science/programs/metabolic-disease-program/publications/mitocarta/mitocarta-in-0
}
\details{
The associated publication is
Calvo, S.E., Klauser, C.R., Mootha, V.K. MitoCarta2.0: an updated inventory of mammalian mitochondrial proteins (2015).  Nucleic Acids Research.

The publication is available from 
\url{http://nar.oxfordjournals.org/content/early/2015/10/07/nar.gkv1003.full}

Please cite the publication if you use the MitoCarta database.
}
\examples{
data(mitocarta2.hg19)
show(mitocarta2.hg19)
}
\keyword{datasets}
