% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTseekerData.R
\name{.onAttach}
\alias{.onAttach}
\title{Data for the MTseeker package examples.}
\usage{
.onAttach(lib, pkgname = "MTseekerData")
}
\arguments{
\item{lib}{the library}

\item{pkgname}{the package name}
}
\value{
nothing, it's a package
}
\description{
Data for the MTseeker package examples.
}
