% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RONKSreads.R
\docType{data}
\name{RONKSreads}
\alias{RONKSreads}
\title{RONKSreads: chrM reads from Renal Oncocytomas and Normal Kidney Samples}
\format{An MAlignmentsList object, which subclasses GAlignmentsList}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA271036/}
}
\usage{
RONKSreads
}
\description{
RONKS == "Renal Oncocytoma, Normal Kidney Sample" matched by patient
This object was produced by applying MTseeker::getMT(BAMs) to the full
exome BAMs, aligned against hg19_rCRSchrM (i.e. GRCh37 with UCSC contigs).
}
\examples{
library(MTseeker)
show(RONKSreads)
}
\keyword{datasets}
